COMMERCE-1453 Move checkoutStep files, constants file, remove export, and suppression
