+++ commerce-checkout-web/bnd.bnd
@@ -1,7 +1,4 @@
 Bundle-Name: Liferay Commerce Checkout Web
 Bundle-SymbolicName: com.liferay.commerce.checkout.web
 Bundle-Version: 1.2.5
-Export-Package:\
-	com.liferay.commerce.checkout.web.constants,\
-	com.liferay.commerce.checkout.web.util
 -dsannotations-options: inherit
