+++ commerce-checkout-web/src/main/java/com/liferay/commerce/checkout/web/internal/display/context/OrderConfirmationCheckoutStepDisplayContext.java
@@ -14,9 +14,9 @@
 
 package com.liferay.commerce.checkout.web.internal.display.context;
 
-import com.liferay.commerce.checkout.web.constants.CommerceCheckoutWebKeys;
 import com.liferay.commerce.checkout.web.internal.display.context.util.CommerceCheckoutRequestHelper;
 import com.liferay.commerce.checkout.web.internal.util.PaymentProcessCommerceCheckoutStep;
+import com.liferay.commerce.constants.CommerceCheckoutWebKeys;
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.model.CommerceOrderPayment;
 import com.liferay.commerce.order.CommerceOrderHttpHelper;
