--- commerce-checkout-web/src/main/resources/com/liferay/commerce/checkout/web/constants/packageinfo
@@ -1,1 +1,0 @@
-version 1.2.0
