--- commerce-checkout-web/src/main/resources/com/liferay/commerce/checkout/web/util/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
