+++ commerce-payment-method-money-order/src/main/resources/META-INF/resources/init.jsp
@@ -23,7 +23,7 @@ taglib uri="http://liferay.com/tld/frontend" prefix="liferay-frontend" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.commerce.checkout.web.constants.CommerceCheckoutWebKeys" %><%@
+<%@ page import="com.liferay.commerce.constants.CommerceCheckoutWebKeys" %><%@
 page import="com.liferay.commerce.payment.method.money.order.internal.configuration.MoneyOrderGroupServiceConfiguration" %><%@
 page import="com.liferay.commerce.payment.method.money.order.internal.display.context.MoneyOrderCheckoutStepDisplayContext" %><%@
 page import="com.liferay.portal.kernel.settings.LocalizedValuesMap" %><%@
