+++ commerce-payment-service/build.gradle
@@ -8,8 +8,6 @@ buildService {
 dependencies {
 	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	compileOnly group: "com.liferay", name: "com.liferay.portal.spring.extender", version: "2.0.0"
-	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.api", version: "9.0.0"
-	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.checkout.web", version: "1.0.0"
 	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.payment.api", version: "1.1.0"
 	compileOnly group: "com.liferay.commerce", name: "com.liferay.commerce.product.api", version: "10.0.0"
 	compileOnly group: "com.liferay.portal", name: "release.portal.api", version: "7.1.1"
@@ -17,4 +15,5 @@ dependencies {
 	compileOnly group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly project(":apps:commerce:commerce-api")
 }
