+++ source-formatter-suppressions.xml
@@ -10,7 +10,6 @@
 	</checkstyle>
 	<source-check>
 		<suppress checks="BNDExportsCheck" files="commerce-application-list/bnd\.bnd" />
-		<suppress checks="BNDExportsCheck" files="commerce-checkout-web/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-initializer-util/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-media-impl/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-price-list-web/bnd\.bnd" />
